+++ oscillator/python/clean.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+set -e -u
+
+. ../../tools/cleaning-tools.sh
+
+rm -rfv ./output/
+
+clean_precice_logs .